implement add brand,category

+++ src/utils/deleteImageFromDisk.js
@@ -1,5 +1,8 @@
 import fs from 'fs-extra';
 const deleteImageFromDiskSync =(filePath) => {
+    if(!fs.existsSync(filePath)){
+        return;
+    } 
     fs.unlinkSync(filePath, (err) => {
         if (err) {
             console.error(`Error deleting the file at ${filePath}:`, err);
